+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.5.0] - 2024-01-18
+
 ### Changed
 * refactor output format of service products-contain-component (GRIF-209)
 * container roots are excluded by default (need to use --include-container-roots)